@@ -17,10 +17,12 @@
   "license": "MIT",
   "devDependencies": {
     "@types/prop-types": "^15.5.3",
-    "@types/react": "^16.4.6",
-    "@types/react-native": "^0.55.26",
-    "react": "16.4.1",
-    "react-native": "0.55.4",
-    "typescript": "^2.9.2"
+    "@types/react": "^16.8.4",
+    "@types/react-native": "^0.57.38",
+    "react": "^16.8.3",
+    "react-native": "^0.58.5",
+    "react-native-windows": "^0.57.0-rc.5",
+    "rxjs": "^6.4.0",
+    "typescript": "^3.3.3333"
   }
 }
